plots/metrics: move to one package, introduce common tests
