+++ tests/func/metrics/metrics/test_show.py
