+++ tests/func/metrics/plots/conftest.py
