+++ tests/func/metrics/plots/test_show.py
@@ -1,4 +1,3 @@
-import csv
 import json
 import logging
 import os
@@ -6,7 +5,6 @@ import shutil
 from collections import OrderedDict
 
 import pytest
-from funcy import first
 
 from dvc.repo import Repo
 from dvc.repo.plots.data import (
@@ -23,25 +21,7 @@ from dvc.repo.plots.template import (
 )
 from dvc.utils.fs import remove
 from dvc.utils.serialize import dump_yaml, dumps_yaml
-
-
-def _write_csv(metric, filename, header=True):
-    with open(filename, "w", newline="") as csvobj:
-        if header:
-            writer = csv.DictWriter(
-                csvobj, fieldnames=list(first(metric).keys())
-            )
-            writer.writeheader()
-            writer.writerows(metric)
-        else:
-            writer = csv.writer(csvobj)
-            for d in metric:
-                assert len(d) == 1
-                writer.writerow(list(d.values()))
-
-
-def _write_json(tmp_dir, metric, filename):
-    tmp_dir.gen(filename, json.dumps(metric, sort_keys=True))
+from tests.func.metrics import _write_csv, _write_json
 
 
 def test_plot_csv_one_column(tmp_dir, scm, dvc, run_copy_metrics):
